requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-cors==0.8.1
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
backoff==2.2.1
beartype==0.15.0
beautifulsoup4==4.13.4
bleach==6.2.0
bokeh==3.7.2
boto3==1.38.8
botocore==1.38.8
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cohere==5.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.10.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fs==2.4.16
fsspec==2025.3.2
ftfy==6.3.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-pubsub==2.29.0
google-cloud-run==0.10.17
google-cloud-secret-manager==2.23.3
google-cloud-vision==3.10.1
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h3==4.2.2
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
instructor==1.2.6
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.1.147
linkify-it-py==2.0.3
litellm==1.44.28
llama-index-core==0.10.68.post1
llama-index-readers-pathway==0.1.3
llama-index-retrievers-pathway==0.1.3
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
msal==1.32.3
multidict==6.4.3
mypy_extensions==1.1.0
narwhals==1.37.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
Office365-REST-Python-Client==2.6.1
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.1
openai==1.77.0
opencv-python==4.11.0.86
openparse==0.5.6
openpyxl==3.1.5
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.18
packaging==23.2
pandas==2.2.3
panel==1.6.3
param==2.2.0
parso==0.8.4
pathway==0.16.4
pdf2image==1.17.0
pdfminer.six==20250416
pexpect==4.9.0
pi_heif==0.22.0
pikepdf==9.7.0
pillow==10.4.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.9.1
pydantic_core==2.23.4
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pypandoc==1.15
pyparsing==3.2.3
pypdf==5.4.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
python-sat==1.8.dev16
pytz==2025.2
pyviz_comms==3.0.4
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==80.3.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlglot==10.6.1
stack-data==0.6.3
surya-ocr==0.4.14
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typer==0.15.3
types-requests==2.32.0.20250328
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
unstructured==0.17.2
unstructured-client==0.28.1
unstructured-inference==0.8.10
unstructured.pytesseract==0.3.15
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.3
xyzservices==2025.4.0
yarl==1.20.0
zipp==3.21.0
